+++ darr/utils.py
@@ -5,12 +5,11 @@ import numpy as np
 from pathlib import Path
 from functools import reduce
 from operator import mul
-import os
 import shutil
 import tempfile as tf
 from contextlib import contextmanager
 
-# believe it or not Python <3.8 does not has such a function
+# believe it or not Python <3.8 does not have such a function
 # and numpy.product returns int32 by default (!) causing disaster
 # when calculating the size of large files
 def product(iterable):